@@ -4,10 +4,10 @@
   "private": true,
   "dependencies": {
     "antd": "^3.2.2",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.1.1"
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "react-router-dom": "^4.3.1",
+    "react-scripts": "1.1.5"
   },
   "scripts": {
     "start": "react-app-rewired start",
