@@ -59,7 +59,7 @@ class App extends Component {
     });
   }
 
-  componentWillMount() {
+  componentDidMount() {
     this.loadCurrentUser();
   }
 
